bump forsuredbapi version to 0.10.1; bump version to 0.10.1; mark FSCursorRecyclerViewAdapter as deprecated; make new BaseFSCursorRecyclerAdapter that does not enforce any particular FSGetApi extension
