+++ forsuredbandroid/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'android-maven'
 apply plugin: 'com.jfrog.bintray'
 
 group = 'com.fsryan.forsuredb'
-version = '0.10.0'
+version = '0.10.1'
 
 android {
     compileSdkVersion 23
@@ -47,7 +47,7 @@ dependencies {
 
     compile 'com.google.guava:guava:19.0'
 
-    compile 'com.fsryan.forsuredb:forsuredbapi:0.10.0'
+    compile 'com.fsryan.forsuredb:forsuredbapi:0.10.1'
     compile 'com.fsryan.forsuredb:sqlitelib:0.4.0'
 
     testCompile 'junit:junit:4.12'
